Update serialization.md
All other includes of properties or other imp things are with grey background but here in this case Utf8Json wasn't there with grey background. It's a minor change but as a reader anyone can figure it out the difference. I hope you will consider this change as it is small but an important Visual chnage.